Google Calendar integration
We are now pulling data from google calendar and there is a very rudimentary
!nextlrr command that shows currently happening and the next future stream.

Hopefully the algorithm for listing these is good enough to handle
the crazyness that happens during the fandraft and any other strange things
that might happen.

Up next, I need to actually make the formatting for the command result better.